requirements.txt
flask==3.0.3
playwright==1.46.0
python-dotenv==1.0.1
gunicorn==22.0.0
